requirements.txt
Flask==0.10.1
Flask-Admin==1.0.7
Flask-SQLAlchemy==1.0
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
SQLAlchemy==0.8.3
Tempita==0.5.1
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.0
argparse==1.2.1
decorator==3.4.0
distribute==0.7.3
itsdangerous==0.23
python-dateutil==2.1
six==1.4.1
wsgiref==0.1.2
